SpatialPooler: extract boosting to a separate file
