+++ src/nupic/research/spatial_pooler.py
@@ -26,7 +26,7 @@ from nupic.bindings.math import (SM32 as SparseMatrix,
                                  SM_01_32_32 as SparseBinaryMatrix,
                                  GetNTAReal,
                                  Random as NupicRandom)
-
+from nupic.research.boosting import Boosting
 
 
 realDType = GetNTAReal()
@@ -94,7 +94,7 @@ class BinaryCorticalColumns(_SparseMatrixCorticalColumnAdapter,
 
 
 
-class SpatialPooler(object):
+class SpatialPooler(Boosting):
   """
   This class implements the spatial pooler. It is in charge of handling the
   relationships between the columns of a region and the inputs bits. The
@@ -239,12 +239,19 @@ class SpatialPooler(object):
       Determines if inputs at the beginning and end of an input dimension should
       be considered neighbors when mapping columns to inputs.
     """
+    columnDimensions = numpy.array(columnDimensions, ndmin=1)
+    numColumns = columnDimensions.prod()
+
+    super(SpatialPooler, self).__init__(
+               numColumns=numColumns,
+               maxBoost=maxBoost,
+               minPctOverlapDutyCycle=minPctOverlapDutyCycle,
+               minPctActiveDutyCycle=minPctActiveDutyCycle
+               )
+
     if (numActiveColumnsPerInhArea == 0 and
         (localAreaDensity == 0 or localAreaDensity > 0.5)):
       raise InvalidSPParamValueError("Inhibition parameters are invalid")
-
-    columnDimensions = numpy.array(columnDimensions, ndmin=1)
-    numColumns = columnDimensions.prod()
 
     if not isinstance(numColumns, (int, long)) or numColumns <= 0:
       raise InvalidSPParamValueError("Invalid number of columns ({})"
@@ -276,10 +283,7 @@ class SpatialPooler(object):
     self._synPermActiveInc = synPermActiveInc
     self._synPermBelowStimulusInc = synPermConnected / 10.0
     self._synPermConnected = synPermConnected
-    self._minPctOverlapDutyCycles = minPctOverlapDutyCycle
-    self._minPctActiveDutyCycles = minPctActiveDutyCycle
     self._dutyCyclePeriod = dutyCyclePeriod
-    self._maxBoost = maxBoost
     self._spVerbosity = spVerbosity
     self._wrapAround = wrapAround
     self._synPermMin = 0.0
@@ -351,14 +355,6 @@ class SpatialPooler(object):
       self._potentialPools.replace(columnIndex, potential.nonzero()[0])
       perm = self._initPermanence(potential, initConnectedPct)
       self._updatePermanencesForColumn(perm, columnIndex, raisePerm=True)
-
-    self._overlapDutyCycles = numpy.zeros(numColumns, dtype=realDType)
-    self._activeDutyCycles = numpy.zeros(numColumns, dtype=realDType)
-    self._minOverlapDutyCycles = numpy.zeros(numColumns,
-                                             dtype=realDType)
-    self._minActiveDutyCycles = numpy.zeros(numColumns,
-                                            dtype=realDType)
-    self._boostFactors = numpy.ones(numColumns, dtype=realDType)
 
     # The inhibition radius determines the size of a column's local
     # neighborhood.  A cortical column must overcome the overlap score of
@@ -480,16 +476,6 @@ class SpatialPooler(object):
     self._dutyCyclePeriod = dutyCyclePeriod
 
 
-  def getMaxBoost(self):
-    """Returns the maximum boost value"""
-    return self._maxBoost
-
-
-  def setMaxBoost(self, maxBoost):
-    """Sets the maximum boost value"""
-    self._maxBoost = maxBoost
-
-
   def getIterationNum(self):
     """Returns the iteration number"""
     return self._iterationNum
@@ -585,90 +571,6 @@ class SpatialPooler(object):
     self._synPermConnected = synPermConnected
 
 
-  def getMinPctOverlapDutyCycles(self):
-    """Returns the minimum tolerated overlaps, given as percent of
-    neighbors overlap score"""
-    return self._minPctOverlapDutyCycles
-
-
-  def setMinPctOverlapDutyCycles(self, minPctOverlapDutyCycles):
-    """Sets the minimum tolerated activity duty cycle, given as percent of
-    neighbors' activity duty cycle"""
-    self._minPctOverlapDutyCycles = minPctOverlapDutyCycles
-
-
-  def getMinPctActiveDutyCycles(self):
-    """Returns the minimum tolerated activity duty cycle, given as percent of
-    neighbors' activity duty cycle"""
-    return self._minPctActiveDutyCycles
-
-
-  def setMinPctActiveDutyCycles(self, minPctActiveDutyCycles):
-    """Sets the minimum tolerated activity duty, given as percent of
-    neighbors' activity duty cycle"""
-    self._minPctActiveDutyCycles = minPctActiveDutyCycles
-
-
-  def getBoostFactors(self, boostFactors):
-    """Returns the boost factors for all columns. 'boostFactors' size must
-    match the number of columns"""
-    boostFactors[:] = self._boostFactors[:]
-
-
-  def setBoostFactors(self, boostFactors):
-    """Sets the boost factors for all columns. 'boostFactors' size must match
-    the number of columns"""
-    self._boostFactors[:] = boostFactors[:]
-
-
-  def getOverlapDutyCycles(self, overlapDutyCycles):
-    """Returns the overlap duty cycles for all columns. 'overlapDutyCycles'
-    size must match the number of columns"""
-    overlapDutyCycles[:] = self._overlapDutyCycles[:]
-
-
-  def setOverlapDutyCycles(self, overlapDutyCycles):
-    """Sets the overlap duty cycles for all columns. 'overlapDutyCycles'
-    size must match the number of columns"""
-    self._overlapDutyCycles[:] = overlapDutyCycles
-
-
-  def getActiveDutyCycles(self, activeDutyCycles):
-    """Returns the activity duty cycles for all columns. 'activeDutyCycles'
-    size must match the number of columns"""
-    activeDutyCycles[:] = self._activeDutyCycles[:]
-
-
-  def setActiveDutyCycles(self, activeDutyCycles):
-    """Sets the activity duty cycles for all columns. 'activeDutyCycles'
-    size must match the number of columns"""
-    self._activeDutyCycles[:] = activeDutyCycles
-
-
-  def getMinOverlapDutyCycles(self, minOverlapDutyCycles):
-    """Returns the minimum overlap duty cycles for all columns.
-    '_minOverlapDutyCycles' size must match the number of columns"""
-    minOverlapDutyCycles[:] = self._minOverlapDutyCycles[:]
-
-
-  def setMinOverlapDutyCycles(self, minOverlapDutyCycles):
-    """Sets the minimum overlap duty cycles for all columns.
-    '_minOverlapDutyCycles' size must match the number of columns"""
-    self._minOverlapDutyCycles[:] = minOverlapDutyCycles[:]
-
-
-  def getMinActiveDutyCycles(self, minActiveDutyCycles):
-    """Returns the minimum activity duty cycles for all columns.
-    '_minActiveDutyCycles' size must match the number of columns"""
-    minActiveDutyCycles[:] = self._minActiveDutyCycles[:]
-
-
-  def setMinActiveDutyCycles(self, minActiveDutyCycles):
-    """Sets the minimum activity duty cycles for all columns.
-    '_minActiveDutyCycles' size must match the number of columns"""
-    self._minActiveDutyCycles = minActiveDutyCycles
-
-
   def getPotential(self, columnIndex, potential):
     """Returns the potential mapping for a given column. 'potential' size
     must match the number of inputs"""
@@ -753,25 +655,21 @@ class SpatialPooler(object):
     self._updateBookeepingVars(learn)
     inputVector = numpy.array(inputVector, dtype=realDType)
     inputVector.reshape(-1)
-    overlaps = self._calculateOverlap(inputVector)
 
-    # Apply boosting when learning is on
-    if learn:
-      boostedOverlaps = self._boostFactors * overlaps
-    else:
-      boostedOverlaps = overlaps
+    overlaps = self._calculateOverlap(inputVector)
+    # Apply boosting
+    boostedOverlaps = self.getBoostedOverlaps(overlaps, learn)
 
     # Apply inhibition to determine the winning columns
     activeColumns = self._inhibitColumns(boostedOverlaps)
 
     if learn:
       self._adaptSynapses(inputVector, activeColumns)
-      self._updateDutyCycles(overlaps, activeColumns)
-      self._bumpUpWeakColumns()
-      self._updateBoostFactors()
-      if self._isUpdateRound():
+      doUpdateRound = self._isUpdateRound()
+      doGlobal = self._globalInhibition or self._inhibitionRadius > self._numInputs
+      self.updateBoosting(overlaps, activeColumns, doUpdateRound, doGlobal)
+      if doUpdateRound:
         self._updateInhibitionRadius()
-        self._updateMinDutyCycles()
 
     activeArray.fill(0)
     activeArray[activeColumns] = 1
@@ -791,97 +689,6 @@ class SpatialPooler(object):
     """
     neverLearned = numpy.where(self._activeDutyCycles == 0)[0]
     activeArray[neverLearned] = 0
-
-
-  def _updateMinDutyCycles(self):
-    """
-    Updates the minimum duty cycles defining normal activity for a column. A
-    column with activity duty cycle below this minimum threshold is boosted.
-    """
-    if self._globalInhibition or self._inhibitionRadius > self._numInputs:
-      self._updateMinDutyCyclesGlobal()
-    else:
-      self._updateMinDutyCyclesLocal()
-
-
-  def _updateMinDutyCyclesGlobal(self):
-    """
-    Updates the minimum duty cycles in a global fashion. Sets the minimum duty
-    cycles for the overlap and activation of all columns to be a percent of the
-    maximum in the region, specified by minPctOverlapDutyCycle and
-    minPctActiveDutyCycle respectively. Functionality it is equivalent to
-    _updateMinDutyCyclesLocal, but this function exploits the globality of the
-    computation to perform it in a straightforward, and more efficient manner.
-    """
-    self._minOverlapDutyCycles.fill(
-        self._minPctOverlapDutyCycles * self._overlapDutyCycles.max()
-      )
-    self._minActiveDutyCycles.fill(
-        self._minPctActiveDutyCycles * self._activeDutyCycles.max()
-      )
-
-
-  def _updateMinDutyCyclesLocal(self):
-    """
-    Updates the minimum duty cycles. The minimum duty cycles are determined
-    locally. Each column's minimum duty cycles are set to be a percent of the
-    maximum duty cycles in the column's neighborhood. Unlike
-    _updateMinDutyCyclesGlobal, here the values can be quite different for
-    different columns.
-    """
-    for i in xrange(self._numColumns):
-      maskNeighbors = numpy.append(i,
-        self._getNeighborsND(i, self._columnDimensions,
-        self._inhibitionRadius))
-      self._minOverlapDutyCycles[i] = (
-        self._overlapDutyCycles[maskNeighbors].max() *
-        self._minPctOverlapDutyCycles
-      )
-      self._minActiveDutyCycles[i] = (
-        self._activeDutyCycles[maskNeighbors].max() *
-        self._minPctActiveDutyCycles
-      )
-
-
-  def _updateDutyCycles(self, overlaps, activeColumns):
-    """
-    Updates the duty cycles for each column. The OVERLAP duty cycle is a moving
-    average of the number of inputs which overlapped with the each column. The
-    ACTIVITY duty cycles is a moving average of the frequency of activation for
-    each column.
-
-    Parameters:
-    ----------------------------
-    @param overlaps:
-                    An array containing the overlap score for each column.
-                    The overlap score for a column is defined as the number
-                    of synapses in a "connected state" (connected synapses)
-                    that are connected to input bits which are turned on.
-    @param activeColumns:
-                    An array containing the indices of the active columns,
-                    the sparse set of columns which survived inhibition
-    """
-    overlapArray = numpy.zeros(self._numColumns, dtype=realDType)
-    activeArray = numpy.zeros(self._numColumns, dtype=realDType)
-    overlapArray[overlaps > 0] = 1
-    activeArray[activeColumns] = 1
-
-    period = self._dutyCyclePeriod
-    if (period > self._iterationNum):
-      period = self._iterationNum
-
-    self._overlapDutyCycles = self._updateDutyCyclesHelper(
-                                self._overlapDutyCycles,
-                                overlapArray,
-                                period
-                              )
-
-    self._activeDutyCycles = self._updateDutyCyclesHelper(
-                                self._activeDutyCycles,
-                                activeArray,
-                                period
-                              )
-
 
 
   def _updateInhibitionRadius(self):
@@ -1025,20 +832,24 @@ class SpatialPooler(object):
       self._updatePermanencesForColumn(perm, columnIndex, raisePerm=True)
 
 
-  def _bumpUpWeakColumns(self):
+  def _calculateOverlap(self, inputVector):
     """
-    This method increases the permanence values of synapses of columns whose
-    activity level has been too low. Such columns are identified by having an
-    overlap duty cycle that drops too much below those of their peers. The
-    permanence values for such columns are increased.
+    This function determines each column's overlap with the current input
+    vector. The overlap of a column is the number of synapses for that column
+    that are connected (permanence value is greater than '_synPermConnected')
+    to input bits which are turned on. Overlap values that are lower than
+    the 'stimulusThreshold' are ignored. The implementation takes advantage of
+    the SparseBinaryMatrix class to perform this calculation efficiently.
+
+    Parameters:
+    ----------------------------
+    @param inputVector: a numpy array of 0's and 1's that comprises the input to
+                    the spatial pooler.
     """
-    weakColumns = numpy.where(self._overlapDutyCycles
-                                < self._minOverlapDutyCycles)[0]
-    for columnIndex in weakColumns:
-      perm = self._permanences[columnIndex].astype(realDType)
-      maskPotential = numpy.where(self._potentialPools[columnIndex] > 0)[0]
-      perm[maskPotential] += self._synPermBelowStimulusInc
-      self._updatePermanencesForColumn(perm, columnIndex, raisePerm=False)
+    overlaps = numpy.zeros(self._numColumns).astype(realDType)
+    self._connectedSynapses.rightVecSumAtNZ_fast(inputVector, overlaps)
+    overlaps[overlaps < self._stimulusThreshold] = 0
+    return overlaps
 
 
   def _raisePermanenceToThreshold(self, perm, mask):
@@ -1270,66 +1081,6 @@ class SpatialPooler(object):
     return potential
 
 
-  @staticmethod
-  def _updateDutyCyclesHelper(dutyCycles, newInput, period):
-    """
-    Updates a duty cycle estimate with a new value. This is a helper
-    function that is used to update several duty cycle variables in
-    the Column class, such as: overlapDutyCucle, activeDutyCycle,
-    minPctDutyCycleBeforeInh, minPctDutyCycleAfterInh, etc. returns
-    the updated duty cycle. Duty cycles are updated according to the following
-    formula:
-
-                  (period - 1)*dutyCycle + newValue
-      dutyCycle := ----------------------------------
-                              period
-
-    Parameters:
-    ----------------------------
-    @param dutyCycles: An array containing one or more duty cycle values that need
-                    to be updated
-    @param newInput: A new numerical value used to update the duty cycle
-    @param period:  The period of the duty cycle
-    """
-    assert(period >= 1)
-    return (dutyCycles * (period -1.0) + newInput) / period
-
-
-  def _updateBoostFactors(self):
-    r"""
-    Update the boost factors for all columns. The boost factors are used to
-    increase the overlap of inactive columns to improve their chances of
-    becoming active. and hence encourage participation of more columns in the
-    learning process. This is a line defined as: y = mx + b boost =
-    (1-maxBoost)/minDuty * dutyCycle + maxFiringBoost. Intuitively this means
-    that columns that have been active enough have a boost factor of 1, meaning
-    their overlap is not boosted. Columns whose active duty cycle drops too much
-    below that of their neighbors are boosted depending on how infrequently they
-    have been active. The more infrequent, the more they are boosted. The exact
-    boost factor is linearly interpolated between the points (dutyCycle:0,
-    boost:maxFiringBoost) and (dutyCycle:minDuty, boost:1.0).
-
-            boostFactor
-                ^
-    maxBoost _  |
-                |\
-                | \
-          1  _  |  \ _ _ _ _ _ _ _
-                |
-                +--------------------> activeDutyCycle
-                   |
-            minActiveDutyCycle
-    """
-
-    mask = numpy.where(self._minActiveDutyCycles > 0)[0]
-    self._boostFactors[mask] = ((1 - self._maxBoost) /
-      self._minActiveDutyCycles[mask] * self._activeDutyCycles[mask]
-        ).astype(realDType) + self._maxBoost
-
-    self._boostFactors[self._activeDutyCycles >
-      self._minActiveDutyCycles] = 1.0
-
-
   def _updateBookeepingVars(self, learn):
     """
     Updates counter instance variables each round.
@@ -1345,26 +1096,6 @@ class SpatialPooler(object):
     self._iterationNum += 1
     if learn:
       self._iterationLearnNum += 1
-
-
-  def _calculateOverlap(self, inputVector):
-    """
-    This function determines each column's overlap with the current input
-    vector. The overlap of a column is the number of synapses for that column
-    that are connected (permanence value is greater than '_synPermConnected')
-    to input bits which are turned on. Overlap values that are lower than
-    the 'stimulusThreshold' are ignored. The implementation takes advantage of
-    the SparseBinaryMatrix class to perform this calculation efficiently.
-
-    Parameters:
-    ----------------------------
-    @param inputVector: a numpy array of 0's and 1's that comprises the input to
-                    the spatial pooler.
-    """
-    overlaps = numpy.zeros(self._numColumns).astype(realDType)
-    self._connectedSynapses.rightVecSumAtNZ_fast(inputVector, overlaps)
-    overlaps[overlaps < self._stimulusThreshold] = 0
-    return overlaps
 
 
   def _calculateOverlapPct(self, overlaps):
